moved analysis and measurement modules into post_treatment
