+++ src/pymodaq/post_treatment/daq_measurement/daq_measurement_GUI.ui
