+++ src/pymodaq/resources/VERSION
@@ -1,1 +1,1 @@
-version = '3.6.3'
+version = '4.0.0'
