+++ tests/daq_utils_test/plotting_test/data_viewers_test/viewer1D_test.py
@@ -8,7 +8,7 @@ from pyqtgraph import ROI
 from unittest import mock
 from collections import OrderedDict
 from pymodaq.utils import daq_utils as utils
-from pymodaq.daq_measurement.daq_measurement_main import DAQ_Measurement
+from pymodaq.post_treatment.daq_measurement.daq_measurement_main import DAQ_Measurement
 from pymodaq.utils.plotting.data_viewers.viewer1D import Viewer1D, Viewer1D_math
 from pymodaq.utils.exceptions import ExpectedError, Expected_1, Expected_2
 from pymodaq.utils.conftests import qtbotskip
